Use correct marker URL for Google china map type
Summary:
We're using marker assets from maps.google.com for GoogleChinaMapType,
which will result in markers not shown because it is not accessible.

Adding a fragment sub-class for china map type, and add
a callback onJavaScriptInit so that we can toggle china mode and use the
correct asset URLs later.

Test Plan:
Added a test case for china map.
1) Simulate china network environment.
2) open china map.
3) before fix: see map loaded but markers not showing up
4) after fix: see map loaded and markers show up as expected.